+++ crates/bench_runner/src/main.rs
@@ -7,7 +7,6 @@ use std::net::SocketAddr;
 use std::path::PathBuf;
 use std::time::{Duration, SystemTime};
 use tokio::net::TcpListener;
-use tokio::runtime::Builder;
 use tokio::sync::oneshot;
 use tokio_postgres::{Client, Config as ClientConfig, NoTls};
